requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web scraping and document processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf2>=3.0.0
python-docx>=0.8.11

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0
